# [baremetal] Harden tinycrypt against fault injection