--- a/library/pkparse.c
+++ b/library/pkparse.c
@@ -987,7 +987,9 @@ static int pk_parse_key_sec1_der( mbedtls_uecc_keypair *keypair,
     {
         ret = uECC_compute_public_key( keypair->private_key,
                                        keypair->public_key );
-        if( ret == 0 )
+        if( ret == UECC_FAULT_DETECTED )
+            return( MBEDTLS_ERR_PLATFORM_FAULT_DETECTED );
+        if( ret != UECC_SUCCESS )
             return( MBEDTLS_ERR_PK_KEY_INVALID_FORMAT );
     }