--- a/library/ssl_cli.c
+++ b/library/ssl_cli.c
@@ -3567,7 +3567,7 @@ static int ssl_out_client_key_exchange_write( mbedtls_ssl_context *ssl,
         == MBEDTLS_KEY_EXCHANGE_ECDH_ECDSA )
 
     {
-        const struct uECC_Curve_t * uecc_curve = uECC_secp256r1();
+        uECC_Curve uecc_curve = uECC_secp256r1();
         ((void) n);
         ((void) ret);
 
@@ -3718,7 +3718,7 @@ static int ssl_out_client_key_exchange_write( mbedtls_ssl_context *ssl,
             == MBEDTLS_KEY_EXCHANGE_ECDHE_PSK )
         {
 #if defined(MBEDTLS_USE_TINYCRYPT)
-            const struct uECC_Curve_t * uecc_curve = uECC_secp256r1();
+            uECC_Curve uecc_curve = uECC_secp256r1();
             ((void) n);
             ((void) ret);