--- a/library/ssl_srv.c
+++ b/library/ssl_srv.c
@@ -3280,7 +3280,7 @@ static int ssl_prepare_server_key_exchange( mbedtls_ssl_context *ssl,
 #endif /* MBEDTLS_KEY_EXCHANGE__WITH_SERVER_SIGNATURE__ENABLED */
 #endif /* MBEDTLS_KEY_EXCHANGE__SOME_PFS__ENABLED */
 #if defined(MBEDTLS_USE_TINYCRYPT)
-    const struct uECC_Curve_t * uecc_curve = uECC_secp256r1();
+    uECC_Curve uecc_curve = uECC_secp256r1();
 #endif
 
     (void) ciphersuite_info; /* unused in some configurations */