… include/mbedtls/config.h
… library/ssl_tls.c
… tests/suites/test_suite_pk.function
… tests/suites/test_suite_pkparse.function
… tests/suites/test_suite_tinycrypt.function
… tinycrypt/ecc.c
… tinycrypt/ecc_dh.c
… tinycrypt/ecc_dsa.c