## Description

**Note to gatekeeper:** this PR makes changes to the TinyCrypt API - see "Migrations" below.

~Dependencies:This PR is based #2918 (merged).~

This PR adds counter-measures against fault injection in ECC computations to our version of TinyCrypt. It also performs some refactorings to simplify implementation of the counter-measures and reduce their impact on code size.

The counter-measures to implement are

- [x] check that the point lies on the curve at the end of the computation
- [x] add integrity check for curve parameters (p, n, G, b)
- [x] add reporting of detected faults
- [x] add double-checking, high Hamming distance and other safety measures where necessary
- [x] document API changes

### Code size impact

Size of the library (sum of all 3 `.a` files, with SSL server enabled) in bytes, as measured with `scripts/baremetal.sh --rom --armc5 | grep total`:

| step | size | diff (prev.) | diff (first) |
| --- | --- | --- | --- |
| Before | 39842 | +0 | +0 |
| After (centralized) point validation | 39746 | -96 | -96 |
| After refactoring curve parameters | 39310 | -436 | -532 |
| After integrity check for curve parameters | 39586 | +276 | -256 |
| After reporting of detected faults | 39682 | +96 | -160 |
| After double-check / flow control | 39774 | +92 | -68 |

Thanks to the refactorings, the PR **saves 68 bytes** of code size despite adding new security checks.

### Impact on speed

The execution speed of various operations was measured on a  `NUCLEO_L073RZ` using Jarno's [test program](https://github.com/ARMmbed/mbedtls-test/tree/master/benchmark/rng_test). The difference, if any, is smaller (0 to 8 ms on a 1.9 second operation) that the measurement noise (up to 10 ms).

The counter-measures added by this PR have **no measurable impact** on speed.

## Status
**READY**

## Migrations

This PR contains API changes, which will require changing all application code using one of the affected APIs directly. The top-level ECDH and ECDSA APIs are affected, so it is very likely that all code using TinyCrypt directly will need changing. Code using TinyCrypt only through the Mbed TLS API (functions starting with `mbedtls_`) is not impacted.

The main changes are a new return value convention, and the removal of the "curve" parameter for functions that had one.

Here's a more detailed and complete list sorted by type of change.

### Change of return value convention.

Some functions used to return `1` for success and `0` for failure, and now return one of `UECC_SUCCESS`, `UECC_FAILURE`, `UECC_FAULT_DETECTED`. Failure to adapt caller code would not result in compile-time or link-time errors, but would cause the code to misbehave at runtime instead, potentially in security-critical ways. Testing might not detect all errors so code review is required.

List of public functions affected:

* `uECC_make_key()`
* `uECC_shared_secret()`
* `uECC_sign()`
* `uECC_verify()`
* `uECC_compute_public_key()`

List of semi-internal functions affected - application code should not need to call these directly unless, so hopefully no changes are needed here, but since they're technically public here's the list anyway:

* `EccPoint_compute_public_key()` - use `uECC_compute_public_key()` instead
* `EccPoint_mult_safer()` - use one of the ECDH or ECDSA APIs instead
* `EccPoint_isZero()` - use `uECC_valid_public_key()` instead

**Actions required:**

* identify all calling points of the affected function in the code
* edit all of them to check the return value as follows: (1) `UECC_SUCCESS` means success, and any other value means failure, but (2) the failing value `UECC_FAULT_DETECTED` might be treated in a special way for logging or reporting purposes, as it denotes a hardware malfunction in a security-critical path, which may be indicative of an active physical attack.

### Change of signature for some functions

The following functions lost a parameter. They were previously accepting a `curve` parameter, but since only a single value for this parameter was supported (namely the return value of `uECC_secp256r1()`), this parameter was useless and has been removed for consistency and to simplify the implementation of some internal checks.

Public functions that lost the `curve` parameter:

* `uECC_make_key()`
* `uECC_shared_secret()`
* `uECC_sign()`
* `uECC_verify()`

* `uECC_curve_private_key_size()`
* `uECC_curve_public_key_size()`
* `uECC_compute_public_key()`
* `uECC_valid_public_key()` - note: you're welcome to call this function, but shouldn't need to, as validity of public keys is already checked by all functions taking them as input anyway.

Semi-internal functions that lost the `curve` parameter - application code should not need to call these directly unless, so hopefully no changes are needed here, but since they're technically public here's the list anyway:

* `EccPoint_compute_public_key()` - use `uECC_compute_public_key()` instead
* `EccPoint_mult_safer()` - use one of the ECDH or ECDSA APIs instead
* `EccPoint_isZero()` - use `uECC_valid_public_key()` instead
* `XYcZ_add()`  - use one of the ECDH or ECDSA APIs instead
* `uECC_valid_point()` - use `uECC_valid_public_key()` instead

**Actions required:**

* edit calling sites of these functions to remove the final `curve` argument;
* if a `curve` variable was used, remove it as well as it's now unused.

### Removal of the `curve` structure and associated types and function

In relation to the previous point (functions loosing their `curve` parameter), the following parts of the API were removed:

* the types `struct uECC_Curve_t` and `uECC_Curve`
* the object `curve_secp256r1`
* the function 

**Actions required:** none in addition to what was already done to adapt to the signature change above.

### Removal of a few internal functions from the public API

The following functions were previously publicly visible while being internal in nature. It is very likely that no application code was using them, but here's the list, with suggestions for replacement:

* `x_side_default()`: use `uECC_valid_public_key()` instead

**Actions required:** hopefully none as you're probably not calling those functions already. Otherwise, adapt code to make use of the suggested higher-level function instead.