@mpg 2019-11-28
The pr-merge job succeeded, and the pr-head job only failed in one instance of the known-flaky "reconnect" `ssl-opt.sh` test - restarting it.