@mpg 2019-11-29
The pr-merge job fully passed, but the pr-head job has two failures that both look spurious - restarting.