@mpg 2019-11-29
@artokin Thanks for your review. I think I've addressed your feedback and this is now ready for you to review again.

@jarlamsa Please also review this when you get the chance. Thanks!