@jarlamsa 2019-12-04
@jarvte could you review the latest commit to this PR, @artokin, please re-review.