@jarlamsa 2019-12-04
With two approvals and a passing CI, no conflicts, and the only prerequisite also ready for merge, this is ready for merge.