@sbutcher-arm 2019-12-05
Failing IAR compiler with: 
```
IAR ANSI C/C++ Compiler V8.20.1.14183/W32 for ARM
   Copyright 1999-2017 IAR Systems AB.
   Network license: 10.10.12.86 (STD)

          diff |= computed[i] ^ reference[i];
          ^
"C:\builds\ws\mbedtls-release-new\src\tinycrypt\ecc.c",163  Error[Pa082]: 
          undefined behavior: the order of volatile accesses is undefined in
          this statement

      diff |= (unsigned char) i ^ 32;
      ^
"C:\builds\ws\mbedtls-release-new\src\tinycrypt\ecc.c",167  Error[Pa082]: 
          undefined behavior: the order of volatile accesses is undefined in
          this statement

          diff |= (left[i] ^ right[i]);
          ^
"C:\builds\ws\mbedtls-release-new\src\tinycrypt\ecc.c",287  Error[Pa082]: 
          undefined behavior: the order of volatile accesses is undefined in
          this statement

Errors: 3
```
Therefore marking as 'needs: work' and removing 'ready to merge'.