@mpg 2019-12-05
@sbutcher-arm Thanks for reporting the error! I hopefully fixed it in a new commit.

@artokin @jarlamsa Could you review the latest commit? Thanks!