@mpg 2019-12-05
The base PR #2918 has been merged, so removing "needs: preceding PR".