@mpg 2019-12-05
(Changed the base twice in order to flush github's cache and have a correct commit list.)