@mpg 2019-12-06
Removing "DO NOT MERGE" and "needs: work" as API changes are now properly documented in the PR description.

The CI is only failing the API/ABI check, which is expected and now documented in the PR description, so removing "needs: ci" as well.

The PR is now only awaiting one more re-review - @jarlamsa or @jarvte could one of you review the commit added since your previous approval? Thanks!