@mpg 2019-12-09
This PR has two approvals, and no conflicts or dependencies.

The CI is only failing the API/ABI check, which is expected as the API is indeed changed. The changes are fully documented in the PR description. All other tests are passing.

So labeling "ready for merge".